# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Jose Madrid Salsa API
  version: 1.0.0
  description: |
    RESTful API for the Jose Madrid Salsa e-commerce platform.
    
    ## Authentication
    Most endpoints require authentication via NextAuth session cookies.
    Admin endpoints require specific permissions based on user roles.
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse.
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "error": "Error message description",
      "code": "ERROR_CODE"
    }
    ```
  contact:
    name: Jose Madrid Salsa Team
    url: https://josemadridsalsa.com
  license:
    name: MIT
servers:
  - url: https://josemadridsalsa.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
  /products:
    get:
      tags:
        - Products
      summary: List all active products
      description: Retrieve a list of all active products with optional filtering
      parameters:
        - name: heatLevel
          in: query
          description: Filter by heat level
          schema:
            type: string
            enum: [MILD, MEDIUM, HOT, EXTRA_HOT, FRUIT, all]
        - name: search
          in: query
          description: Search products by name or description
          schema:
            type: string
        - name: featured
          in: query
          description: Filter featured products only
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        featuredImage:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        heatLevel:
          type: string
          enum: [MILD, MEDIUM, HOT, EXTRA_HOT, FRUIT]
        sku:
          type: string
        inventory:
          type: integer
        isFeatured:
          type: boolean
        ingredients:
          type: array
          items:
            type: string
        searchKeywords:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
...